Combine entry/root and AppProviders for App.tsx entrypoint
