+++ frontends/mit-open/src/App.tsx
@@ -1,13 +1,29 @@
 import React, { StrictMode } from "react"
+import { createRoot } from "react-dom/client"
 import { HelmetProvider } from "react-helmet-async"
-
-import { RouterProvider } from "react-router-dom"
+import { createBrowserRouter, RouterProvider } from "react-router-dom"
 import type { RouterProviderProps } from "react-router-dom"
-
+import invariant from "tiny-invariant"
+import * as Sentry from "@sentry/react"
 import { QueryClientProvider, QueryClient } from "@tanstack/react-query"
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools"
-import { ThemeProvider } from "ol-design"
 import { Provider as NiceModalProvider } from "@ebay/nice-modal-react"
+import { ThemeProvider } from "ol-design"
+import { createQueryClient } from "./libs/react-query"
+import routes from "./routes"
+
+Sentry.init({
+  dsn: window.SETTINGS.sentry_dsn,
+  release: window.SETTINGS.release_version,
+  environment: window.SETTINGS.environment,
+})
+
+const container = document.getElementById("app-container")
+invariant(container, "Could not find container element")
+
+const root = createRoot(container)
+const router = createBrowserRouter(routes)
+const queryClient = createQueryClient()
 
 interface AppProps {
   router: RouterProviderProps["router"]
@@ -17,7 +33,7 @@ interface AppProps {
 /**
  * Renders child with Router, QueryClientProvider, and other such context provides.
  */
-const AppProviders: React.FC<AppProps> = ({ router, queryClient }) => {
+const App: React.FC<AppProps> = ({ router, queryClient }) => {
   return (
     <StrictMode>
       <ThemeProvider>
@@ -37,4 +53,4 @@ const AppProviders: React.FC<AppProps> = ({ router, queryClient }) => {
   )
 }
 
-export { AppProviders }
+root.render(<App queryClient={queryClient} router={router} />)
